MakerController & MakerService

+++ src/main/java/com/bankingeen/backofficeservice/model/dto/ScenarioDTO.java
@@ -1,7 +1,7 @@
 package com.bankingeen.backofficeservice.model.dto;
 
 public record ScenarioDTO(
-        int scenarioId
+        Long scenarioId
         ,String scenarioName
         ,String tableName
         , String filterQuery
